a bit of hopefully nondestructive optimisation in biquad

+++ src/ay.h
@@ -60,9 +60,9 @@ public:
         int tone_src = this->tons >> ch;        // tone source
         int noise_ena_l = mixer >> (ch + 3);    // noise enable
         int noise_gen_op = this->noiv;          // noise source
-        int mix = ((tone_ena_l | tone_src) & (noise_ena_l | noise_gen_op)) & 1;
+        float mix = ((tone_ena_l | tone_src) & (noise_ena_l | noise_gen_op)) & 1;
 
-        float result = (float)mix * amp[mode_l ? this->envv : ((this->ayr[8 + ch] & 0x0f))];
+        float result = mix * amp[mode_l ? this->envv : ((this->ayr[8 + ch] & 0x0f))];
         return result;
     }
